refactor(gc): split collector internals by responsibility
